Move delete action to Edit form
